+++ src/controllers/hiring.controller.ts
@@ -4,19 +4,14 @@ import { IApiReply } from "../interfaces/api.interface";
 class HiringController {
   public async makeSoftDeleteHiringReq(hiringId: string): Promise<IApiReply> {
     const appEnv = import.meta.env.VITE_APP_ENV;
-    const baseUrl =
-      appEnv === "local" ? "/api/v1" : import.meta.env.VITE_API_BASE_URL;
+    const baseUrl = appEnv === "local" ? "/api/v1" : import.meta.env.VITE_API_BASE_URL;
     let url = `${baseUrl}/hiring/soft-delete/${hiringId}`;
 
     try {
       const appController = new CWPBApiController();
 
       const rawReply = await appController.DELETE(url);
-      const reply = await appController.getSafeReply(
-        rawReply,
-        url,
-        appController.DELETE
-      );
+      const reply = await appController.getSafeReply(rawReply, url, appController.DELETE.bind(appController));
 
       return reply;
     } catch {
@@ -29,23 +24,16 @@ class HiringController {
     }
   }
 
-  public async makePermanentDeleteHiringReq(
-    hiringId: string
-  ): Promise<IApiReply> {
+  public async makePermanentDeleteHiringReq(hiringId: string): Promise<IApiReply> {
     const appEnv = import.meta.env.VITE_APP_ENV;
-    const baseUrl =
-      appEnv === "local" ? "/api/v1" : import.meta.env.VITE_API_BASE_URL;
+    const baseUrl = appEnv === "local" ? "/api/v1" : import.meta.env.VITE_API_BASE_URL;
     let url = `${baseUrl}/hiring/delete/${hiringId}`;
 
     try {
       const appController = new CWPBApiController();
 
       const rawReply = await appController.DELETE(url);
-      const reply = await appController.getSafeReply(
-        rawReply,
-        url,
-        appController.DELETE
-      );
+      const reply = await appController.getSafeReply(rawReply, url, appController.DELETE.bind(appController));
 
       return reply;
     } catch {
@@ -60,22 +48,18 @@ class HiringController {
 
   public async makeGetHiringRecordsReq(): Promise<IApiReply> {
     const appEnv = import.meta.env.VITE_APP_ENV;
-    const baseUrl =
-      appEnv === "local" ? "/api/v1" : import.meta.env.VITE_API_BASE_URL;
+    const baseUrl = appEnv === "local" ? "/api/v1" : import.meta.env.VITE_API_BASE_URL;
     let url = `${baseUrl}/hiring/all`;
 
     try {
       const appController = new CWPBApiController();
 
       const rawReply = await appController.GET(url);
-      const reply = await appController.getSafeReply(
-        rawReply,
-        url,
-        appController.GET
-      );
+      const reply = await appController.getSafeReply(rawReply, url, appController.GET.bind(appController));
 
       return reply;
-    } catch {
+    } catch (error) {
+      console.log(error, "error");
       const reply = {
         status: ApiStatus.TIMEOUT,
         message: "Connection broked, please try again later",