+++ src/interfaces/index.ts
@@ -14,6 +14,7 @@ export type InputChange = React.ChangeEvent<HTMLInputElement | HTMLTextAreaEleme
 export type AlertType = "error" | "info" | "warning" | "success";
 
 export type QueryString = Record<string, string | number>;
+export type ApiPayload = Record<string, any>;
 export type POSTCallbackFn = (url: string, qs?: QueryString, payload?: Record<string, any>) => Promise<Response>;
 export type GETCallbackFn = (url: string, qs?: QueryString) => Promise<Response>;